+++ src/entities/album/model/types.tsx
@@ -3,10 +3,4 @@ export type Album = {
   id: number;
   title: string;
 };
-export type Photo = {
-  albumId: number;
-  id: number;
-  title: string;
-  url: string;
-  thumbnailUrl: string;
-};
+
