+++ src/entities/album/ui/AlbumsWithPhoto.module.css
@@ -5,24 +5,3 @@
 .albumTitle {
   margin-bottom: 10px;
 }
-
-.photosGrid {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 10px;
-}
-
-.photoItem {
-  width: 100px;
-  text-align: center;
-}
-
-.photoThumbnail {
-  width: 100%;
-  border-radius: 4px;
-}
-
-.photoTitle {
-  font-size: 12px;
-  margin-top: 4px;
-}
