+++ src/entities/album/ui/AlbumsWithPhoto.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { albumsApi } from "../api/albumsApi";
 import type { Album } from "../../../types/AlbumType";
+import PhotoCard from "../../photo/ui/PhotoCard";
 import styles from "./AlbumsWithPhoto.module.css";
 interface AlbumWithPhotosProps {
   album: Album;
@@ -20,17 +21,7 @@ const AlbumWithPhotos: React.FC<AlbumWithPhotosProps> = ({ album }) => {
       <h3 className={styles.albumTitle}>{album.title}</h3>
       <div className={styles.photosGrid}>
         {photos.map((photo) => (
-          <div key={photo.id} className={styles.photoItem}>
-            <img
-              src={photo.url.replace(
-                /(https:\/\/via\.placeholder\.com)(\/(\d+)\/([^/]+))/,
-                "https://placehold.co$2/FFF"
-              )}
-              alt={photo.title}
-              className={styles.photoThumbnail}
-            />
-            <p className={styles.photoTitle}>{photo.title}</p>
-          </div>
+          <PhotoCard photo={photo} />
         ))}
       </div>
     </div>
