+++ src/entities/photo/model/types.ts
@@ -1,0 +1,7 @@
+export type Photo = {
+  albumId: number;
+  id: number;
+  title: string;
+  url: string;
+  thumbnailUrl: string;
+};
