+++ src/entities/photo/ui/PhotoCard.module.css
@@ -1,0 +1,20 @@
+.photosGrid {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 10px;
+}
+
+.photoItem {
+  width: 100px;
+  text-align: center;
+}
+
+.photoThumbnail {
+  width: 100%;
+  border-radius: 4px;
+}
+
+.photoTitle {
+  font-size: 12px;
+  margin-top: 4px;
+}
